There is no way there are MInput issues left
Shifted MInput disabler schedule duration thing to an engine update hook, and have it just be a clock that constantly sets to disable when +ve, and enable when -ve. it also ticks down. and like when it reaches the min value it stops changing MInput so it doesn't affect other mods COUGH imguihelper COUGH